+++ CHANGELOG.md
@@ -6,5 +6,6 @@ v1.1.0 - March 16, 2023
 #### Highlights
 - feature: Add postOnly flag to order responses
 - feature: Add ioc flag to order responses
-- feature: Add ioc and postOnly parameters to create margin/spot limit 
+- feature: Add ioc and postOnly parameters to create margin/spot limit
+- feature: Add endpoints for get user custom fee and user fee for some market 
 - doc: Added an example of using websocket